pyproject.toml
[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
include = ["app", "tests", "scripts"]
exclude = ["**/__pycache__", "**/.pytest_cache", ".venv"]
typeCheckingMode = "basic"
reportMissingImports = true
reportUnusedImport = true
reportUnusedVariable = true

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | alembic
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "__pycache__", "alembic"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/alembic/*"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auth-service"
version = "1.0.0"
description = "Multi-app authentication and authorization service"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "bcrypt==4.1.1",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "alembic==1.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "httpx==0.25.2",
    "black==23.11.0",
    "flake8==6.1.0",
    "isort==5.12.0",
    "mypy==1.7.1",
]
